requirements.txt
annotated-types==0.5.0 ; python_version >= "3.9" and python_version < "3.13"
cachetools==5.3.1 ; python_version >= "3.9" and python_version < "3.13"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "3.13"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.9" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.13" and platform_system == "Windows"
google-api-core==2.11.1 ; python_version >= "3.9" and python_version < "3.13"
google-auth==2.23.0 ; python_version >= "3.9" and python_version < "3.13"
google-cloud-core==2.3.3 ; python_version >= "3.9" and python_version < "3.13"
google-cloud-storage==2.10.0 ; python_version >= "3.9" and python_version < "3.13"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "3.13"
google-resumable-media==2.6.0 ; python_version >= "3.9" and python_version < "3.13"
googleapis-common-protos==1.60.0 ; python_version >= "3.9" and python_version < "3.13"
greenlet==2.0.2 ; python_version >= "3.9" and python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
idna==3.4 ; python_version >= "3.9" and python_version < "3.13"
numpy==1.25.2 ; python_version >= "3.9" and python_version < "3.13"
pandas==2.1.0 ; python_version >= "3.9" and python_version < "3.13"
protobuf==4.24.3 ; python_version >= "3.9" and python_version < "3.13"
psycopg2-binary==2.9.7 ; python_version >= "3.9" and python_version < "3.13"
pyasn1-modules==0.3.0 ; python_version >= "3.9" and python_version < "3.13"
pyasn1==0.5.0 ; python_version >= "3.9" and python_version < "3.13"
pydantic-core==2.6.3 ; python_version >= "3.9" and python_version < "3.13"
pydantic==2.3.0 ; python_version >= "3.9" and python_version < "3.13"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.13"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "3.13"
pytz==2023.3 ; python_version >= "3.9" and python_version < "3.13"
requests==2.31.0 ; python_version >= "3.9" and python_version < "3.13"
rsa==4.9 ; python_version >= "3.9" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.13"
sqlalchemy==2.0.20 ; python_version >= "3.9" and python_version < "3.13"
tabulate==0.9.0 ; python_version >= "3.9" and python_version < "3.13"
toml==0.10.2 ; python_version >= "3.9" and python_version < "3.13"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "3.13"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "3.13"
urllib3==1.26.16 ; python_version >= "3.9" and python_version < "3.13"
